# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Blog API",
    "description": "\n## 博客系统 API\n\n一个功能完整的现代化博客系统后端 API，支持 Markdown/MDX 内容管理和 GitOps 自动化同步。\n\n### 核心特性\n\n- 🔐 **用户认证**：基于 JWT 的安全认证系统\n- 📝 **内容管理**：支持 Markdown/MDX 格式的文章创建和编辑\n- 🎨 **三种渲染模式**：后端渲染、前端 SSR、前端 CSR\n- 📁 **媒体管理**：图片、视频等媒体文件的上传和管理\n- 🔄 **GitOps 同步**：Git 仓库与数据库的双向自动化同步\n- 🔗 **Webhook 集成**：支持 GitHub Webhook 自动触发同步\n\n### 技术栈\n\n- **框架**：FastAPI + SQLModel + PostgreSQL\n- **认证**：JWT + OAuth2\n- **内容处理**：markdown-it-py + MDX\n- **异步支持**：asyncio + aiofiles\n\n### 快速开始\n\n1. 注册账号：`POST /api/v1/users/register`\n2. 登录获取 token：`POST /api/v1/users/login`\n3. 使用 token 访问受保护的接口：`Authorization: Bearer <token>`\n\n### 文档\n\n- [Swagger UI](/docs) - 交互式 API 文档\n- [Scalar UI](/scalar) - 现代化 API 文档\n- [OpenAPI JSON](/openapi.json) - OpenAPI 规范文件\n    ",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/posts/{post_type}/admin/posts": {
      "get": {
        "tags": [
          "posts"
        ],
        "summary": "获取指定板块的文章列表（管理后台）",
        "operationId": "listPostsByTypeAdmin",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "post_type",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PostType",
              "description": "板块类型"
            },
            "description": "板块类型"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/PostStatus"
                },
                {
                  "type": "null"
                }
              ],
              "description": "文章状态（draft/published/archived）",
              "title": "Status"
            },
            "description": "文章状态（draft/published/archived）"
          },
          {
            "name": "category_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "description": "分类ID",
              "title": "Category Id"
            },
            "description": "分类ID"
          },
          {
            "name": "tag_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "description": "标签ID",
              "title": "Tag Id"
            },
            "description": "标签ID"
          },
          {
            "name": "author_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "description": "作者ID",
              "title": "Author Id"
            },
            "description": "作者ID"
          },
          {
            "name": "is_featured",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "是否为推荐文章",
              "title": "Is Featured"
            },
            "description": "是否为推荐文章"
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "搜索关键词（标题、内容）",
              "title": "Search"
            },
            "description": "搜索关键词（标题、内容）"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "每页数量",
              "default": 20,
              "title": "Limit"
            },
            "description": "每页数量"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "偏移量",
              "default": 0,
              "title": "Offset"
            },
            "description": "偏移量"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_PostShortResponse_"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/posts/admin/posts": {
      "get": {
        "tags": [
          "posts"
        ],
        "summary": "获取所有文章列表（管理后台，跨板块）",
        "operationId": "listAllPostsAdmin",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/PostStatus"
                },
                {
                  "type": "null"
                }
              ],
              "description": "文章状态（draft/published/archived）",
              "title": "Status"
            },
            "description": "文章状态（draft/published/archived）"
          },
          {
            "name": "category_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "description": "分类ID",
              "title": "Category Id"
            },
            "description": "分类ID"
          },
          {
            "name": "tag_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "description": "标签ID",
              "title": "Tag Id"
            },
            "description": "标签ID"
          },
          {
            "name": "author_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "description": "作者ID",
              "title": "Author Id"
            },
            "description": "作者ID"
          },
          {
            "name": "is_featured",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "是否为推荐文章",
              "title": "Is Featured"
            },
            "description": "是否为推荐文章"
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "搜索关键词（标题、内容）",
              "title": "Search"
            },
            "description": "搜索关键词（标题、内容）"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "每页数量",
              "default": 20,
              "title": "Limit"
            },
            "description": "每页数量"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "偏移量",
              "default": 0,
              "title": "Offset"
            },
            "description": "偏移量"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_PostShortResponse_"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/posts/me": {
      "get": {
        "tags": [
          "posts"
        ],
        "summary": "获取当前用户的文章列表",
        "operationId": "getMyPosts",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/PostStatus"
                },
                {
                  "type": "null"
                }
              ],
              "description": "文章状态（draft/published/archived）",
              "title": "Status"
            },
            "description": "文章状态（draft/published/archived）"
          },
          {
            "name": "category_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "description": "分类ID",
              "title": "Category Id"
            },
            "description": "分类ID"
          },
          {
            "name": "tag_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "description": "标签ID",
              "title": "Tag Id"
            },
            "description": "标签ID"
          },
          {
            "name": "author_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "description": "作者ID",
              "title": "Author Id"
            },
            "description": "作者ID"
          },
          {
            "name": "is_featured",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "是否为推荐文章",
              "title": "Is Featured"
            },
            "description": "是否为推荐文章"
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "搜索关键词（标题、内容）",
              "title": "Search"
            },
            "description": "搜索关键词（标题、内容）"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "每页数量",
              "default": 20,
              "title": "Limit"
            },
            "description": "每页数量"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "偏移量",
              "default": 0,
              "title": "Offset"
            },
            "description": "偏移量"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_PostShortResponse_"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CategoryResponse": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "title": "Name"
          },
          "slug": {
            "type": "string",
            "maxLength": 100,
            "title": "Slug"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "default": ""
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true
          },
          "sort_order": {
            "type": "integer",
            "title": "Sort Order",
            "default": 0
          },
          "icon_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Icon Id"
          },
          "icon_preset": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Icon Preset"
          },
          "post_type": {
            "$ref": "#/components/schemas/PostType",
            "default": "article"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "name",
          "slug",
          "id"
        ],
        "title": "CategoryResponse"
      },
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message",
          "timestamp",
          "request_id"
        ],
        "title": "ErrorDetail",
        "description": "错误详情模型"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "description": "统一错误响应模型"
      },
...
      "Page_PostShortResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PostShortResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page"
          },
          "size": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Size"
          },
          "pages": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Pages"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "size",
          "pages"
        ],
        "title": "Page[PostShortResponse]"
      },
...
      "PostShortResponse": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200,
            "title": "Title"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "post_type": {
            "$ref": "#/components/schemas/PostType",
            "default": "article"
          },
          "status": {
            "$ref": "#/components/schemas/PostStatus",
            "default": "draft"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "excerpt": {
            "type": "string",
            "title": "Excerpt"
          },
          "reading_time": {
            "type": "integer",
            "title": "Reading Time"
          },
          "view_count": {
            "type": "integer",
            "title": "View Count"
          },
          "like_count": {
            "type": "integer",
            "title": "Like Count"
          },
          "comment_count": {
            "type": "integer",
            "title": "Comment Count",
            "default": 0
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "published_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "author_id": {
            "type": "string",
            "format": "uuid",
            "title": "Author Id"
          },
          "cover_media_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Media Id"
          },
          "author": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "category": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CategoryResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "cover_media": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MediaFileResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "tags": {
            "items": {
              "$ref": "#/components/schemas/TagResponse"
            },
            "type": "array",
            "title": "Tags",
            "default": []
          },
          "git_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Git Hash"
          },
          "source_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Path"
          }
        },
        "type": "object",
        "required": [
          "title",
          "slug",
          "id",
          "excerpt",
          "reading_time",
          "view_count",
          "like_count",
          "created_at",
          "updated_at",
          "author_id"
        ],
        "title": "PostShortResponse",
        "description": "用于列表展示的精简响应，规避 N+1 风险"
      },
      "PostStatus": {
        "type": "string",
        "enum": [
          "draft",
          "published",
          "archived"
        ],
        "title": "PostStatus",
        "description": "文章状态枚举"
      },
      "PostType": {
        "type": "string",
        "enum": [
          "article",
          "idea"
        ],
        "title": "PostType",
        "description": "内容类型枚举"
      },
...
      "UserResponse": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 50,
            "minLength": 3,
            "title": "Username",
            "description": "用户名"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email",
            "description": "邮箱"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "description": "是否激活"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole",
            "description": "用户角色",
            "default": "user"
          },
          "full_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name",
            "description": "全名"
          },
          "bio": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Bio",
            "description": "用户简介"
          },
          "avatar": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar",
            "description": "用户头像 URL"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "用户 ID"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "创建时间"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "更新时间"
          },
          "last_login": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Login",
            "description": "最后登录时间"
          }
        },
        "type": "object",
        "required": [
          "username",
          "email",
          "is_active",
          "id",
          "created_at",
          "updated_at"
        ],
        "title": "UserResponse",
        "description": "用户响应模型（不包含密码）",
        "examples": [
          {
            "created_at": "2025-12-02T10:00:00",
            "email": "alice@example.com",
            "id": "123e4567-e89b-12d3-a456-426614174000",
            "is_active": true,
            "role": "user",
            "updated_at": "2025-12-02T10:00:00",
            "username": "alice"
          }
        ]
      },
      "UserRole": {
        "type": "string",
        "enum": [
          "user",
          "admin",
          "superadmin"
        ],
        "title": "UserRole"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/users/login"
          }
        }
      }
    }
  },
...
}